backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-docx==1.1.0
PyMuPDF==1.23.8
pdfplumber==0.10.3
transformers==4.35.2
torch==2.1.1
numpy==1.24.3
requests==2.31.0
Pillow==10.1.0